+++ libp2p/security/base_transport.py
@@ -3,14 +3,14 @@ from typing import (
     Callable,
 )
 
+from libp2p.abc import (
+    ISecureTransport,
+)
 from libp2p.crypto.keys import (
     KeyPair,
 )
 from libp2p.peer.id import (
     ID,
-)
-from libp2p.security.secure_transport_interface import (
-    ISecureTransport,
 )
