+++ libp2p/transport/typing.py
@@ -8,15 +8,13 @@ from typing import (
 
 from libp2p.abc import (
     IMuxedConn,
+    ISecureTransport,
 )
 from libp2p.custom_types import (
     TProtocol,
 )
 from libp2p.io.abc import (
     ReadWriteCloser,
-)
-from libp2p.security.secure_transport_interface import (
-    ISecureTransport,
 )
 
 THandler = Callable[[ReadWriteCloser], Awaitable[None]]
